tests(TSExtPaths, TSNonExtPaths): Add TSExtPaths and TSNonExt Pathas
